thunderstore-api: allow null values for team donation link
The backend uses Django's cursed null=True for saving the donation
link, so submitting empty string will fail on save (or even earlier by
the validation done by serializer) unless it is cast to null. At least
earlier a design principle has been that API does what it does, and the
clients should adjust their behaviour to match the API. It's also seems
to be cleaner to do the casting here.