Implement package detail utils
Implement two functions responsible for getting public or private
package listings from the same endpoint.

This is required for users who should be able to see for instance rejected
package listings.

The function for getting public package listings is to be used in the
SSR loader function, and the function for getting private package listings
is to be used in the clientLoader function.