Disable package endpoints that are not community-scoped
Content moderators currently reject packages by marking the
PackageListing object rejected. PackageListings are community-scoped.

Some of the endpoints are package scoped, i.e. they are not related to
any communities or listings. These endpoints therefore currently have
no way of telling if a package has been rejected or not. In worst case,
the rejected packages can contain malware, so under no circumstances do
we want to show such packages on the website where unsuspecting users
can be tricked to download the mods.

Therefore the pages of the website that use such endpoints are disabled
until we figure out how to address the problem. Since the old website
doesn't have unscoped endpoints, from feature parity perspective the
new one doesn't need to have them either.

Footgun warning: the website still uses some endpoints that are not
community-scoped, such as "get README for this version of the package".
This is fine as long as the page where the data is shown takes measures
to ensure the content can be shown by querying the package listing
endpoint too and showing 404 error if the package is rejected.