Add package version support for getPackageListingDetails Dapper method
PackageListing as a concept is version-agnostic, but to ensure we don't
serve pages for rejected package versions, they need to be
community-scoped, as rejection is tied to PackageListing object.

This commit adds the support to Dapper and listingUtils, the actual
usage will be done in a separate commit.

Since by default a PackageListing shows the latest package version, the
version argument remains optional.