packageEdit: use two-phase package listing fetch
Attempt to fetch package listing data first as unauthenticated user in
hopes of hitting cache. If this results in 404, retry as authenticated
user in hopes of having permissions to view a rejected package listing.

Since we block rendering be awaiting permissions and listing, we might
as well trigger filter loading before we know whether user can access
the page or not. A little bit on overfetch is preferred to awaiting in
phases.